Refactor as requested on PR review
